requirements.txt
Django
pillow
apscheduler
uuid
requests
django_otp
xlwt
django-phonenumber-field
phonenumbers